pyproject.toml
[project]
name = "skillpod"
version = "0.1.0.dev0"
description = "Project-scoped, reproducible skill dependency manager for AI coding agents."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Daniel Hsieh" }]
keywords = ["ai", "agents", "skills", "dependency-manager", "claude", "codex"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.12",
    "pydantic>=2.6",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.scripts]
skillpod = "skillpod.cli:app"

[project.urls]
Homepage = "https://github.com/danielhsieh/skillpod-cli"

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skillpod"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B", "SIM"]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["src/skillpod"]
